Update patch set 11
Patch Set 11:

This is ok for me, but is a bit odd because the commit message raises the expectation this patch removes the event everywhere, which is now does not any more. Please update the commit message and add FIXMEs to the lines you added in PS10 and 11.

Patch-set: 11
Label: Verified=0