Update patch set 1
Patch Set 1:

This is a different algorithm, based on a different assumption (that what the user enters is always a URL), with different tests. This is not an alternative that makes the other patch obsolete. Both do make sense, depending on the strings users are pasting. Both patches should be merged are rebased on top of each other. Both should probably be rebased on top of I77b88de to make the JSCS and JSHint changes obsolete.

Patch-set: 1