Update patch set 2
Patch Set 2:

I'm sorry, but why was this merged within an hour over the weekend, with nobody else having a chance to look at this? This change appears to be wrong. It's not linked to a ticket. The commit message does not explain why this was done. It does not even explain well what this patch actually does (*where* the change is made).

The current code guarantees this is an item ID:

https://github.com/wikimedia/mediawiki-extensions-Wikibase/blob/master/client/includes/Specials/SpecialEntityUsage.php#L61

Either this code is wrong and every mention of "items" must be replaced with generic entities and a generic entity ID parser. Or This must be reverted (plus more mentions of "entity" must be replaced with "item").

Patch-set: 2