Update patch set 1
Patch Set 1:

Daniel, I still do not understand why you need this hook and can not use the "WikibaseChangeNotification" hook that does the exact same thing, just triggered 4 instead of 2 times.

I even believe using the old "WikibaseChange" is actually wrong, because it is not triggered on create and modify, only on delete and undelete. I do not expect an "on delete and undelete" hook to be named "WikibaseChange". This is at least a very bad name.

Patch-set: 1