Update patch set 1
Patch Set 1:

@Thiemo: WikibaseChangeNotification is fired on the repo, that's (probably) not what we want. Though it would also be an option - the handler would then have to send some sort of message to a service that does the actual update work on WDQS.

I didn't look closely at when this hook is fired. If it's only fired "sometimes", it should be renamed or fixed or removed.

I'm perfectly good with removing this hook, I just wnated to mention that a use case for it came up just a few days ago.

Patch-set: 1