Update patch set 6
Patch Set 6: Code-Review-1

Please update this patch as Siebrand already suggested: Remove the GENDER keyword when it contains a single parameter only. Adding this to English messages is confusing and pointless and does not help translators. You are free to use the GENDER keyword as often as you want in non-English languages on translatewiki.net, in places where it actually makes sense. These might even be different places in different languages.

Patch-set: 6
Label: Code-Review=-1