Update patch set 7
Patch Set 7: Code-Review-1

(3 comments)

"Discuss" what?

This patch is sitting here for almost a year. Multiple people spend their precious time on reviewing this, trying to understand it's intention, raising questions and giving advice to help this patch not getting stuck, but approved and merged instead. I kindly ask you to respect these peoples time and efforts.

We already established (and I hope we agree on that) that no translator is blocked by this patch. Every translator is free to use GENDER either with a $… parameter when a message does provide such a parameter with a user name, or with no first parameter to target the current user. I believe https://www.mediawiki.org/wiki/Localisation#.E2.80.A6on_user_names_via_GENDER does confirm this.

https://www.mediawiki.org/wiki/Localisation#.E2.80.A6on_user_names_via_GENDER requires the corresponding qqq message to mention a $… parameter when it contains a user name that can be used in GENDER. I believe this is already the case, since this patch never touched any qqq message.

https://www.mediawiki.org/wiki/Localisation#.E2.80.A6on_user_names_via_GENDER goes further and suggests to use GENDER in the English message, even when the English message does not need it. This gives translators a very obvious signal that they can and should make use of GENDER. This is especially helpful when there is a $… parameter with a user name. Not so much when there is none. And not so much if it is not obvious that a word or phrase is gender-specific. For example, this patch originally contained phrases like "you {{GENDER:|agree}}". I still do not understand why "agree" was in the GENDER tag, and not "you" or "you agree" together. This may make sense in certain languages, but not in the languages the reviewers of this patch know. So no, I'm very sorry, but after a year it should become obvious that the users you asked for a review here are not able to approve such changes.

Neither the commit message explains the intent of this patch, nor is there a Phabricator ticket assigned that could have helped in resolving all this. You are right, Gerrit is not for discussing such things. This is what wiki pages and Phabricator are for, and what a commit message on Gerrit should link to, to not run into the exact situation we have right now.

Patch-set: 7
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1