Update patch set 1
Patch Set 1: Code-Review+1 Verified+1

Looks OK! Only thing that came to my mind while looking at this that as this is a test relevant for all components, not only for Lib, maybe it could be worth having it in a "top-level" tests directory, ie. in tests/phpunit (or so), instead of lib/tests/phpunit.
On the other hand, Lib is not really a "component" so the way it is now also makes some sense.

Patch-set: 1
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=+1
Label: Code-Review=+1