Skip Julian to Gregorian date conversion for extreme, unsupported dates
This changes the current status quo that was introduced just recently. New
behavior: Non-Gregorian dates we currently can not convert to Gregorian
simply fall back to Gregorian, with the day and month unconverted. This is
based on the assumption that a Julian day out of that range does not make
much sense and was most probably meant to be a Gregorian date.

I believe this behavior is much better than "dropping" day and month and
always setting it to January 1st.

We can not change the precision value from 11 (day) to 9 (year). This must
reflect the actual value as it is stored in the knowledge base.

We can not use https://www.w3.org/TR/xmlschema11-2/#gYear because it is
restricted to 4 digits.

Bug: T146356
Change-Id: I27e7616f568f5188bde1d8302fe8effdaa414260