Add extreme Julian date test to TimeRdfBuilderTest
This documents the current status quo, which is: Non-Gregorian dates we
currently can not convert to Gregorian fall back to the year, January 1st.

We can not change the precision value from 11 (day) to 9 (year). This must
reflect the actual value as it is stored in the knowledge base.

We can not use https://www.w3.org/TR/xmlschema11-2/#gYear because it is
restricted to 4 digits.

Bug: T146356
Change-Id: I27e7616f568f5188bde1d8302fe8effdaa414260