Update patch set 1
Patch Set 1: Code-Review+1

It looks like this patch attempts to "fix" much more that what was reported in T147127. I believe most of this is not necessary and can never happen. But it can't hurt much to have these additional null checks.

Which line fixes the actual bug? Do we need to backport this somehow?

Patch-set: 1