Update patch set 4
Patch Set 4: Code-Review-1

I'm sorry, but how is this an improvement? All this patch does is removing parts of a feature, but what is the benefit?

I expect a suggester to always find at least something, even if what I copy and paste is a partial URL. Example: If I copy and paste "ttps://de.wikipedia.org" because I missed to highlight the first character, I do not expect the suggester to fail - because of a single character. :-( A character that does not even matter, because all the suggester needs to know is there. :-(

Maybe you missed the fact that this subdomain search only triggers if the first search a few lines above returned nothing?

Patch-set: 4
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1