Cleaner JulianDateTimeValueCleaner implementation
This is mostly, but not exclusively refactoring. This also adds a test
and fixes an actual bug: The base class never checked if it is ok to
convert something that is not a Gregorian date. Note that this edge case
is currently not tested because of the "$calendar ===
TimeValue::CALENDAR_GREGORIAN" in the tests. But I did not wanted to make
this patch bigger.

Bug: T146356
Change-Id: Ib23582019a567310dfb87fbd1b51c17999bacc0f