Do not wrap question mark icon to the next line
Note that the toolbar already had a nowrap rule applied, but it was only
applied to the three "save", "delete" and "cancel" buttons before the question
mark icon. The icon was the only "dangling" thing that was able to wrap to the
next line.

With this patch the icon will still overlap with other elements. This is not
new and not introduced by this patch because the three toolbar buttons already
had the same issue. This should be fixed in later patches.

Bug: T146691
Change-Id: I7610eae122ec70deab028e3ec4508a7f23b52ad5