Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

I don't think this is the best way to solve this. A setting listing "entity types that support both labels and descriptions" is extremely specific for this single (set of) special pages.

We are trying to get rid of assumption like "all entities have labels" and "labels and descriptions belong together" for a long time now. Instead we do have interfaces for this: LabelProvider and DescriptionProvider.

The special page can iterate all entity types and check if an entity type supports one or both of these interfaces. This should be fairly easy to do with the existing settings.

Patch-set: 2
Label: Code-Review=-1