Use FORMAT_PLAIN for data access output
Our output format is basically wikitext escaped plain text, except
for URLs, these are not wikitext escaped.

Please note that the tests in DataAccessSnakFormatterOutputFormatTest
verify that the output doesn't change.

We do this so that we can alter the FORMAT_WIKI output
without having to worry about breaking any consumers
of the {{#property:…}} parser function or of the equivalent
Lua functionality.

Bug: T147580
Change-Id: If81faa21e75b308124de7083305f84251c898a34