Remove jquery.effects.blind dependency from tagadata
This patch removes a minor animation effect that is not used anyway.
Main reason for this patch is to get rid of this jQuery dependency.
There is a lot more in the tagadata code I would like to remove but I
wanted this patch to focus on a single detail.

Change-Id: Iff459f28d9f5faa6265ed5345722eb6492b429b1