Update patch set 2
Patch Set 2:

I would love to help with this patch, but I don't understand what the benefit of this new configuration is and how it helps dealing with the wb_entity_per_page technical depth.

This is what I would do:

The special pages iterate all known entity types. We do not touch this code. But the special pages must check two things: #1 If an entity type does have labels (or descriptions). #2 If these labels (or descriptions) are available via the service the special pages use.

How to solve #1: Inject the EntityFactory. Call newEmpty. Check for "instanceof LabelsProvider" (or "instanceof DescriptionsProvider").

How to solve #2: Don't. For now just make sure getEntitiesWithoutTerm returns an empty array for entity types that are not indexed in the wb_entity_per_page table. We deal with this problem when users complain.

Patch-set: 2