Update patch set 2
Patch Set 2:

I had the impression this patch is about the technical depth, not about performance. T140891#2479927 is a comment in a ticket unrelated to performance issues. It also does not sound like a PM decision to me. Was this discussed in story time? Do the special pages really cause so much trouble that we must disable them?

Patch-set: 2