Update patch set 2
Patch Set 2: Code-Review-1

(4 comments)

The code is very obvious, but I'm not happy with many of the naming decisions in here.

This is mostly a copy of the existing DispatchingSnakFormatter, which already does "the same" and dispatches based on property type (and other information).

I believe this class is to generic ("overengenering"). I believe we will never use this class again, only for the described use case. If this is true, we can easily hard-code the knowledge necessary to resolve the described use case: Hard code the property type (it's only one, right?) and hard code what should happen in the two cases. Name the class accordingly. I believe this will be more transparent and much easier to understand when looking at this code later.

Patch-set: 2