Update patch set 2
Patch Set 2:

Ok. Can we please create a separate ticket for this? And clarify what problem this solves? I can see that the SQL query is very critical, but it's part of a special page that's barely used anyway. So again, what problem do we solve by disabling it for Items?

I can see a setting "indexTermsForNumericEntityTypes". This is null by default, which means all numeric entity types (all with Int32EntityIds) are indexed in the terms table. In production we can set it to [ 'property' ]. This will disable indexing for Items.

The special page will then simply show empty results because it can not search for empty labels and descriptions any more.

A dedicated setting to disable a special page still seems odd. At least it should have an obvious name then. Example: "entityTypesWithEmptyTerms", again with a null default, and set to [ 'property' ] on production.

Patch-set: 2