Update patch set 2
Patch Set 2:

Let me try again: This patch adds a setting that does nothing but disabling a specific entity type on a specific special page. This is more a hack than anything else, and will most probably be removed in the near future. The name of the setting should reflect this very obviously.

The default for the setting should be "null", in my opinion. It should default to what it was before: the configured entity types. Otherwise nobody will find this setting when trying to use this special page with a custom entity type.

Patch-set: 2