Update patch set 2
Patch Set 2:

Looks good to me too!

The approach looking exactly like the one for EntityLookup [1] and considering that there are many other services that would need the same multiplexing service makes me wonder whether there should be a generic ServiceMultiplexer (or something along those lines) that also has params for the local service and a map of foreign services.

We talked briefly about this in person before and I guess it depends on how many other services like this there are and how many cases where it wouldn't work exactly like this.

[1] https://github.com/wmde/WikibaseDataModelServices/pull/150

Patch-set: 2
Label: Code-Review=0