Update patch set 4
Patch Set 4: Code-Review+1

I believe this patch is ok (thanks a lot for submitting it, as well as linking to a Phabricator ticket). But it needs a test. Preferably a test case in EditEntityActionTest, or at least a manual test.

This code also needs a lot of refactoring. This should be done in later patches.

Patch-set: 4