Update patch set 4
Patch Set 4: Code-Review-1

Oh, there is a problem. Lets assume you want to revert three revisions. The current revision number is 5, and you want to revert to revision 2. You click the button. You are explaining your revert. You click the final button. But in the meantime somebody did an other edit and created revision number 6. With this patch this sixth revision will also be reverted. This did not happened before this patch.

See my refactoring in I05ee6b2. This should avoid this effect. Please try to apply the relevant change to your patch, or tell me if it is ok for you if I merge both patches into one (I would append to your patch).

Patch-set: 4
Label: Code-Review=-1