Introduce DataAccessSnakFormatterFactory::newRichWikitextSnakFormatter
This is a first step towards implementing new functionality
that uses FORMAT_WIKI, instead of escaped plain text.

This is not yet used anywhere, but we need this to a) test
the new output format and b) to introduce the new Lua functionality
and parser functions.

Note: We might want to consider renaming EscapingSnakFormatter to
WrappedSnakFormatter or something similar.

Bug: T142940
Change-Id: Id96d335bcbc207f0f7265b62b8483ed715931b3e