Introduce missing "not-a-language" error message
Trivial to test on monolingual values.

This message is also used in TermValidatorFactory, which is used in
ChangeOpLabel and so on, which is used in FingerprintChangeOpFactory,
which is used in all kinds of special pages. But these special pages
all have a dedicated check and error message for invalid language codes:
"wikibase-wikibaserepopage-invalid-langcode". This makes
"wikibase-validator-not-a-language" a fallback that's currently not
seen by the user (except for monolingual, see above).

Change-Id: Iede9569575ac57a89663d10c3aa53671000443e7