Add parameters to "wikibase-error-ui-no-external-page" error message
Note that this message is used twice. The effect on Special:SetSiteLink
is quite easy to test. The other instance is more complicated:

In the sitelinks UI, try to add a new sitelink (or edit an existing one)
with a valid site ID but a page name that does not exist on that site.
Click "save". You will see two error messages: the first one is meant to
be generic, but localized. The second is not localized, but gives details.
This patch does make the first, localized message more specific and
actionable.

This requires a release of the WikibaseJavaScriptApi component! Otherwise
the user will see unparsed parameters ("$2"). See I044a28c.

Change-Id: I99d8d799dbce54503fa7f13d175464d856e6ed2f