Update patch set 2
Patch Set 2: Code-Review-1

(4 comments)

I still disagree with hard-coding knowledge about specific entity types here. This is not necessary. We could merge this patch almost as it is when we disagree that this is a migration patch and we are going to continue working on this, especially lifting the temporary limitations this chain of patches introduces.

Questions and suggestions inside. The -1 is to get your attention and because of the incomplete PHPDoc.

Patch-set: 2
Label: Code-Review=-1