Update patch set 3
Patch Set 3: Code-Review+1

Pair reviewed with Jonas. Manually tested:

* This is copy-pasted from the repo settings. Nothing obviously wrong with that.
* Confirmed this does not change the output of the {{#property:…}} parser function, when used on the client.
* There is no wbformatvalue on the client.
* wbformatvalue is on the repo, but not affected by this change.

We can not confirm the patch does what it's title says ("link external ids when used in wikitext"). We can not see how this would happen.

For some reason the linked ticked talks about EntityId formatting, not external ID formatting. Is this a mistake?

We still do not understand what the intention of this change is.

Changing the output of the existing {{#property:…}} function would be a breaking change. I believe we agree we will not do this.

Patch-set: 3