Update patch set 1
Patch Set 1:

(5 comments)

A chain of 6 small, self-contained patches is not "giant". The first two are refactoring and make the following patches much easier to review. All other literally depend on each other, either because they are touching the same code multiple times, or because the functionality of a later patch would be broken without the previous one.

If you need more explanation please ask me on IRC or somewhere else. I don't think Gerrit comments are a good way to resolve something like this.

Patch-set: 1
Label: Verified=0