Update patch set 5
Patch Set 5: Code-Review-1

(2 comments)

An EntityTitleLookup must return the Title of the page containing the entity, not of any page that is somehow connected to the entity. A separate interface should be defined for that if need be. 

Even if the two interfaces have the exact same signature, they have different contracts.

Patch-set: 5
Label: Code-Review=-1