Update patch set 7
Patch Set 7: Code-Review+1

I'm not entirely happy with the message keys, because they are constructed from the name of the widget + "add", but they do not say *what* they add. The existing message keys are "wikibase-addreference", for example.

Do you mind if we change the message keys to "wikibase-addproperty" and "wikibase-addstatement"?

Patch-set: 7