Update patch set 4
Patch Set 4:

This patch is meant as a migration path, to be as small as possible, self-contained, easy and fast to review. I will not touch other files in this patch.

What you are proposing makes sense, but must be done in later patches and can not be a reason for a -1. EntityNamespaceLookup currently contains 3 public methods. I believe getEntityNamespaces can not be part of a "lookup" interface. We should review the usages of this method first.

Furthermore, even when EntityNamespaceLookup is an interface, we can not immediately get rid of the RepoEntityTitleLookup interface I'm introducing here. There are users of this combined interface. This must be refactored, too. So I see a chain of three to five patches to split, refactor and merge all this properly.

Patch-set: 4