Update patch set 5
Patch Set 5:

(1 comment)

I find the panic in your comments here not helpful.

The interface is literally a general "get Title from EntityId lookup". Nothing more. If used on a repo, it typically returns the title of the page holding the entity. If used on a client, it may return an interwiki link to the repo, or a link to the corresponding local page, depending on the implementation.

There is nothing wrong with that. Neither the namespace of the interface (it's in lib, not in repo), nor it's name, nor it's documentation disallows these use cases.

The more specific interface you are talking about is repo exclusive and must be in repo. This is exactly what I'm doing in the next patch.

There are a few users assuming the title they get is the page holding the entity, and others doing nothing but rendering the title as a link or URL. This mix of use cases did existed before (you are referring to this as "de-facto contract"). This is not introduced in this patch and can not be a reason for a -1.

What especially makes me sad is that your -1 does not explain how we can proceed here.

Patch-set: 5
Label: Verified=0