Update patch set 7
Patch Set 7: Code-Review-1

> There are a few users assuming the title they get is the page holding the entity, and others doing nothing but rendering the title as a link or URL. This mix of use cases did existed before (you are referring to this as "de-facto contract"). This is not introduced in this patch and can not be a reason for a -1.

Indeed. So far, there was an (sadly undocumented) assumption that the service returns the page holding the entity. Some code relies on this, some code doesn't care.

You now propose to add an implementation that violates that expectation, which means implementations of the interface are not interchangeable, which should be avoided if sanely possible (without generics, sometimes we can't). You can argue that the expectation is unfounded, because the documentation of the interface does not specify the semantics of the Title it returns. Ok, then we should fix this, by allowing code that expects to get the title of the entity data to express/enforce that expectation, by using an interface that provides this guarantee.

> What especially makes me sad is that your -1 does not explain how we can proceed here.

Code that needs the title of the EntityPage needs to type-hint against an interface that guarantees this. So we need two interfaces, for local-page-somhow-representing-entity and page-that-holds-entity-data. Both may actually be relevant on the client, too: page-that-holds-entity-data can be a Title holding an external link to the repo. Note that both may be useful on both repo and client.

I'm undecided if the more restrictive interface that guarantees that the Title is the page that contains the entity data should be separate/incompatible to the more general interface, or if it rather be an empty extension of that interface, purely a marker.

Patch-set: 7
Label: Code-Review=-1