Update patch set 8
Patch Set 8:

Thinking about it, I'm kind of torn about the representation of an empty list. We shouldn't wrap an empty rendering of a value (or even statement), but should we also not wrap an empty list of statements?

Representing the empty list by an empty string seems convenient. I see no practical issues. But it still kind of feels wrong, like a mistake of category.

Jonas has pointed out that we may want to include meta-data on an empty list. Maybe it wasn't really empty, it just has no element that actually produced a rendering.

But that seems a bit hypothetical to me. I'm still good with using the empty string for empty lists. 

I'm not so happy about the special case for one-element lists. That seems like asking for trouble, especially in the case of css styles.

Patch-set: 8
Label: Verified=0