Update patch set 6
Patch Set 6:

I'm okay with changing to something else but this seems to be a common practice when there is nothing to show. See for example: https://en.wikipedia.org/w/api.php?action=query&generator=pageswithprop&gpwppropname=notock 
(Just remove the "k" at the end and it will return with lots of results)
I can ask an API guru to tell me about it.

Patch-set: 6