Update patch set 7
Patch Set 7:

> So far, there was an (sadly undocumented) assumption that the service returns the page holding the entity.

I disagree, and already explained that. What you say makes sense for an interface in the repo directory, but this is in the lib directory. It is *not* used as you describe. See the lines in WikibaseValueFormatterBuilders I'm touching in this patch.

What you want is an other interface in repo that currently doesn't exist. This must be done in later patches.

> Code that needs the title of the EntityPage needs to type-hint against an interface that guarantees this.

This is exactly what I do in the next patch. I believe I already said this.

> Note that both may be useful on both repo and client.

Yes, that may be the case. Again, see the next patch in this chain, which does exactly that.

Your concerns sound like what I'm doing here (introducing a new implementation for an existing interface) instantly *breaks* all existing users of the interface. This is not the case.

Patch-set: 7
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0