Update patch set 9
Patch Set 9: Code-Review+2

I did not understood what Marius did in the later patch sets. Therefore the -2, so nothing bad gets merged accidentally. As far as I can see this is all reverted now. Is this correct?

The one and only guarantee we want to give is that the output is always a single DOM element (or nothing).

Therefore we must wrap multiple elements in an other span. Otherwise some outputs would be "<span>2001</span>, <span>2002</span>". This would have multiple issues that can all be avoided by wrapping it in something like <span class="wb-statements">…</span>.

If the output is a single DOM element anyway, there is no need to do this. The contrary: It will cause unnecessary confusion if a single statement is wrapped in something that claims to be a list, but is none. This becomes very obvious the moment we support other list renderings. An <ul> with a single element is bad. Instead the element itself should be returned, not wrapped in anything.

I'm merging this for two reasons: To get this chain of patches moving, and because the current implementation does not know if the list elements returned by the SnakFormatter are valid DOM elements.

We must fix this the moment we support <ul> rendering, possibly earlier.

Patch-set: 9
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2