Update patch set 2
Patch Set 2: Code-Review+2

I think this should go in.

I can't think of anything this could break. It should however be announced as a "significant" change according to our Stable Interface Policy, but not a "breaking" change.

DEPLOY: post a heads-up about this to commons-l and commons village pump.

Patch-set: 2
Label: Code-Review=+2