Update patch set 2
Patch Set 2:

Oh, I just saw this old comment of mine on phabricator:

> The idea behind restricting to Properties is: Without wb_entity_per_page, joining against the wb_terms table is inefficient. It would still work for Properties though, because there aren't that many of them.

So instead of hardcoding Items and Properties, we could just hardcode properties and fail for anything else. This isn't going to work for anything else in production anyway. That would not be nice for 3rd parties, but it would be a pragmatic solution to the need at hand, and it would simplify the code.

Patch-set: 2