Update patch set 2
Patch Set 2:

I suggest the following compromise:

Inject the entity-type/prefix mapping as a constructor argument. Let the bootstrap code worry about how to construct it. As a first step, I think it is OK to hard-code Item and Property in the bootstrap code. A smarter mechanism can be added in a follow up.

In any case, this class should not be concerned with how to determine the ID prefixes. It just needs them, so it asks for them in the constructor.

Patch-set: 2