Update patch set 8
Patch Set 8:

About the version - the question is what the client needs. I.e. is there a case when the client would need to distinguish pre-change format from post-change format. With addition, the case is clear - if the client can't process new stuff, they need to know it. With removal, it may be that some client expects all components to be there, but it is arguable how legit such client is. I would bump it anyway just out of caution, but if you decide otherwise it's not a huge deal in this case.

Patch-set: 8