WIP: Update for API error i18n
See Iae0e2ce3.

This patch doesn't actually do anything, because I don't understand what
your code is doing well enough to try. I'm submitting it mainly so
Ib0c9d5a1 can depend on it. Please amend it to actually do what is
needed.

The biggest change is that UsageException is deprecated. The
replacement, ApiUsageException, holds the error messages (plural!) in a
StatusValue object. These are *probably* going to be ApiMessage
instances. The point is that all your unit tests that refer to
UsageException are going to need updating.

I'd hope your ApiErrorReporter class can mostly become a wrapper around
the corresponding core methods. Obviously a lot of the formatting is
different; perhaps you could check whether $apiModule->getErrorFormatter()
is an instance of ApiErrorFormatter_BackCompat to determine whether to
use your existing code or to take advantage of the new core methods.

Depends-On: Iae0e2ce3bd42dd4776a9779664086119ac188412
Change-Id: I26c052fd4867ee17ab9a21b1cc904054df6f8c39