Disable remove buttons (on reference snaks) during save
Edit a reference somehow and save. You will see that the top remove
button (for the entire reference) is disabled during the API request,
but the individual buttons for the snaks are not.

I was looking for a Phabricator ticket but could not find one. Maybe
this was on IRC or on the contact page?

Change-Id: I47db2660bfa467e825ea65ab686ae8121fae0513