Destroy "remove reference" button in non-edit mode
From what I remember Adrian did this on purpose. His idea was that it
should be possible to remove a reference without being in edit mode.
That's why the remove button is always there, but we had it hidden via
CSS.

Multiple user scripts and gadgets make the toolbar visible, which makes
the remove button visible. We do not want this. Instead we should destroy
the remove button along with the add button.

This makes https://github.com/Wikidata/primarysources/pull/128/files
obsolete.

Bug: T148060
Change-Id: I611085db309c99e3d40d7c6fe8c8e8ddd9924e88