Update patch set 1
Patch Set 1: Code-Review+2

The relevant effect of this patch is that the "rich wikitext" output runs much more tests.

I would like to suggest an other way of arranging this code. Have only one provider with the input and *both* expected outputs. Both test functions use the same provider, and each uses an other expected output. This way all the relevant information is very close together in each test case. The code to construct the inputs is not duplicated any more. But some outputs are "duplicated" (not really, because they are not the same anyway because of the <span>s).

I believe we need to rework this code the moment the <span>s have some attributes.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2