Update patch set 4
Patch Set 4:

Note that PS4 changes the behaviour in cases of entity id from some other repository compared to what it used to do in previous patch sets.
Now it does NOT throw the exception but makes the lookup return null/false as the entity was not found in the repository.

Patch-set: 4
Label: Verified=0