Refactor DataAccessSnakFormatterFactory
This is split from If613ac0 to make it easier to review. The goal is to
reduce duplication. Otherwise there would be the same if-else in multiple
places in the code base, just to call these two methods that don't do
anything but turning their name into a string parameter.

Bug: T113955
Change-Id: I1308fb883af64d8e9df3d27dba0b044f19f40cc5