More specific return type in LanguageFallbackLabelDescriptionLookup
See I386ab28 why this is needed. A caller that *knows* that he is calling
a fallback lookup should be allowed to know that he will always get a
fallback object. Giving this guarantee is very helpful and saves useless
instanceof checks. See the other patch for an example.

A caller that does not know he is calling a fallback lookup type hints
against the interface and is unaffected by this change.

Bug: T135045
Change-Id: I17702ed6ed8b2be802e3f384fd9e638542d8265e