Update patch set 2
Patch Set 2:

@Thiemo @Daniel:
this patch indeed depends on Ib999422e876

As said in that patch, I will move the check if the entity id belongs to the "right" repository down to WikiPageMetaDataAccessor/Lookup. This might make things more clear, you could see this once I am done with moving things over.

I agree that adding this generic EntityIdParser here added some confusion. The only reason I added it this is that I wanted to drop the prefix and then still have the instance of EntityId based on the serialization without the prefix. Daniel is right that for DB queries it should be enough to only use the string of the prefix-less id serialization. I only wanted to still have an EntityId instance, as db queries also need a namespace of the entity etc but I could still use the prefixed version of the entity id to pass it to EntityNamespaceLookup etc. I am not sure how clear it is when I describe it, code would speak better to itself :)

As now all stripping etc would happen in WikiPageMetaDataLookup and not in WikiPageEntityRevisionLookup, I believe issues you Thiemo point out here would be more clear and evident.

Please let me adjust the patch and let's discuss it again. Thanks for those valuable comments!

Patch-set: 2
Label: Verified=0