Update patch set 5
Patch Set 5: Code-Review+1

Code looks perfectly fine to me. The relevant effect of this patch is that it will add language fallback capabilities to the undo diff (and only to this very specific diff). This is not mentioned in the commit message, unfortunately.

I have not tested this myself yet.

I want to give others a chance to see this. Poke me if it's not merged next week.

Patch-set: 5
Label: Code-Review=+1