Update patch set 5
Patch Set 5:

Switched to perform the check in WikiPageEntityMetaDataLookup. This seems reasonable.

the only difference with this change compared with PS4, when it comes to impact on WikiPageRevisionEntityLookup is that when WikiPageRevisionEntityLookup::getEntityRevision gets an entity id from other repository, and non-zero revision id, it will now throw BadRevisionException, while before it would have returned null.
But I think that actually the current behaviour (exception) is more consistent.

Patch-set: 5
Label: Verified=0