Update patch set 2
Patch Set 2: Code-Review-1

-1 as it is already based on outdated master.
If this is a move we want to do, I'd simply re-run phpcbf to have all occurrences fixed.

Actually I think it would be nice to dump MW CS dependency to 0.7.2 or so but switching to that version results in few more issues (really few, less than 10) than just array syntax, so I'd rather do it separately.

Patch-set: 2