BinaryOptionDispatchingSnakFormatter fails on unknown properties
This turned out to be a blocker for what we are developing in
ArticlePlaceholder. The entire Special:AboutTopic fails hard when an
Item contains a Property that got deleted. This should not happen.
There is even a proper message rendered. Nothing "TODO" here.

This patch also fixes the broken namespace(s).

Bug: T147580
Change-Id: I57b2c74a4a5cdcb66e116ab3cbd1e0b5e9ce12f6