Rename EntityTitleLookup to EntityTitleStoreLookup
We do need a generic "EntityTitleLookup" interface, but the name is
occupied. I'm doing the first step here: Rename the existing class and
make it's contract much, much more clear. That's all this patch does.

This is an alternative for I96889cd and I866457e. I96889cd and I866457e
can be rebased on top of this. I just did not wanted to do this before we
agree on something.

Personally I believe this is *worse* than I96889cd and I866457e alone.
Look at the massive amount of lines this patch touches for no reason.
Most of this will be reverted when the other two patches are rebased on
top of this. Plus this patch will break PropertySuggester (two times!),
while the other two will not. But if this helps getting the other two
unstuck, I don't mind.

Bug: T148626
Change-Id: Ief1e17291f302aa1c2d80642a4fed6b802160ad3