Update patch set 6
Patch Set 6: Code-Review+1

It was a bit more complicated to make the test fail with the old code and succeed with the new. Please have a look. +2 from my side, if this is ok for you.

Patch-set: 6
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1